backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Werkzeug==3.0.1
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
xgboost==2.0.3
openpyxl==3.1.2
reportlab==4.0.8
python-dotenv==1.0.0